requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
pyjwt==2.8.0
python-dotenv==1.0.0